fix: cast non-sharded modules to bf16 in fully_shard, restore fp32 FSDP precision
Root cause: on_train_begin casts everything to precision_fsdp (float32) for
optimizer precision. With per-block FSDP2 sharding, only CausalDiTBlocks get
FSDP's automatic dtype handling. Non-sharded modules (patch_embedding, embeddings,
head, audio) stay fp32, causing Conv3d dtype mismatch with bf16 inputs.

Fix: after FSDP shards the blocks, explicitly cast non-sharded modules to
compute dtype (bf16). This matches FSDP1's behavior where the unshard hook
auto-casts parameters to param_dtype during forward.

SF config restored to precision_fsdp="float32" for proper fp32 optimizer states.

+++ fastgen/configs/experiments/OmniAvatar/config_sf.py
@@ -99,7 +99,7 @@ def create_config():
 
     # Precision
     config.model.precision = "bfloat16"
-    config.model.precision_fsdp = "bfloat16"  # Must match model precision to avoid dtype mismatch in Conv3d
+    config.model.precision_fsdp = "float32"  # fp32 optimizer precision; non-sharded modules cast to bf16 in fully_shard()
 
     # Input shape: 512x512 @ 81 frames → latent [16, 21, 64, 64]
     config.model.input_shape = [16, 21, 64, 64]
